+++ app/Http/Controllers/ReviewsController.php
@@ -15,8 +15,8 @@ class ReviewsController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth:api', ['except' => ['index', 'show']]);
-        $this->middleware('admin', ['except' => ['index', 'show']]);
+        $this->middleware('auth:api', ['except' => ['index', 'show', 'post']]);
+        $this->middleware('admin', ['except' => ['index', 'show', 'post']]);
     }
     
     /**